requirements.txt
flask
flask-wtf
wtforms
email-validator
flask-sqlalchemy
flask-login
pillow
itsdangerous
flask-mail
